requirements.txt
boto3==1.26.35
click==8.1.3
pandas==1.5.2
snowflake-connector-python==2.9.0
